requirements.txt
discord.py>=2.0.0
pymongo>=4.0.0